Modified Collection and Element classes to accept \GuzzleHttp\Message\Response as a constructor arg for #54.
This is a backward-compatible modification to Acquia\Rest\Collection and Acquia\Rest\Element so that the constructor may be used interchangeably with Guzzle 3.x and higher. Since guzzle isn't strictly required for the acquia/acquia-sdk-php-rest sub-package, I've added the old and present versions of Guzzle as composer suggestions, and removed the requirement on Guzzle 3.

The constructor for Acquia\Rest\Collection is no longer strongly typed, but will throw an InvalidArgumentException if an unusual object is passed to it.